% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mef.R
\name{Build_AverageScTable}
\alias{Build_AverageScTable}
\title{Build of an average tabmoy from dtoto with several usms and seeds}
\usage{
Build_AverageScTable(dtoto, keysc)
}
\description{
Build of an average tabmoy from dtoto with several usms and seeds
}
