% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltoto.R
\name{build_dtoto}
\alias{build_dtoto}
\title{This function builds dtoto}
\usage{
build_dtoto(ls_toto)
}
\arguments{
\item{ls_toto}{list of all toto files of the same experiment}
}
\value{
A dataframe. Each row of this list correspond to an IDusm
}
\description{
This function builds dtoto
}
