% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mef.R
\name{build_dtoto_binary_zip1}
\alias{build_dtoto_binary_zip1}
\title{This function ...calcul indices par espece pour un single zip file / unitary process}
\usage{
build_dtoto_binary_zip1(
  zfile,
  DOYdeb = 60,
  DOYScoupe = c(187, 229, 282, 334),
  ls_parsd = c("Len", "Lfeuille", "phyllochron", "Vmax2", "ELmax", "PPtreshh"),
  cote = 16,
  nblignes = 16,
  opt_saveindices = F
)
}
\description{
This function ...calcul indices par espece pour un single zip file / unitary process
}
