% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mef.R
\name{build_dtoto_binarySim}
\alias{build_dtoto_binarySim}
\title{This function ...calcul indices par espece}
\usage{
build_dtoto_binarySim(ltoto, dtotoBase, ls_toto_paquet, DOYdeb, DOYScoupe)
}
\description{
This function ...calcul indices par espece
}
