% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moy_simtoto.R
\name{build_simmoy}
\alias{build_simmoy}
\title{This function groups all the mean of the output variables in one data frame}
\usage{
build_simmoy(ltoto, lsusm, esp = NA, optSD = F)
}
\arguments{
\item{ltoto}{A large list of three data frames (seeds 0, 1 and 2)}

\item{lsusm}{names(ltoto)}

\item{esp}{Species}

\item{optSD}{Optional parameter.Returns standard deviation of the sum of the individuals}
}
\value{
A data frame grouping all the outputs variables and their mean values for each DOY
}
\description{
This function groups all the mean of the output variables in one data frame
}
