% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_indices.R
\name{Calc_Beta_coeff}
\alias{Calc_Beta_coeff}
\title{Calculate Beta competition coefficient for intra-specific competition}
\usage{
Calc_Beta_coeff(x)
}
\arguments{
\item{x}{A dtoto data.frame of the pure species containing Ytot, densite, nbplt and surfsolref vectors}
}
\value{
A model fit result with "a","b" and "beta" values
}
\description{
Calculate Beta competition coefficient for intra-specific competition
}
\examples{
dtoto <- tabtoto_compet
sp_dtoto <- split(dtoto, dtoto$keysc)
key <-"88-1 Fix2-nonFixSimTest Lusignan30IrrN2 -"
keypur <- "88-1 nonFixSimTest-nonFixSimTest Lusignan30IrrN2 -"
dat <- rbind(sp_dtoto[[key]], sp_dtoto[[keypur]])
pur1 <- dat[dat$densite2 == 0., ]
Calc_Beta_coeff(pur1)
}
