% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_indices.R
\name{Calc_CESE_diag}
\alias{Calc_CESE_diag}
\title{Calculate Loreau's CEi and SEi coefficient for a diagonal}
\usage{
Calc_CESE_diag(diag)
}
\description{
Calculate Loreau's CEi and SEi coefficient for a diagonal
}
