% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_indices.R
\name{Calc_CompetitionIntensity}
\alias{Calc_CompetitionIntensity}
\title{Calculate Competition Intensity coefficients}
\usage{
Calc_CompetitionIntensity(YEsp1, densite1, iso_ind)
}
\description{
Calculate Competition Intensity coefficients
}
