% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_indices.R
\name{Calc_Eij_coefficients}
\alias{Calc_Eij_coefficients}
\title{Calculate Eij coefficients}
\usage{
Calc_Eij_coefficients(res1, res2, parameters1, parameters2)
}
\description{
Calculate Eij coefficients
}
