% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_indices.R
\name{Calc_Gamma_coeffesp12}
\alias{Calc_Gamma_coeffesp12}
\title{Calculate Gamma competition coefficient for inter-specific competition}
\usage{
Calc_Gamma_coeffesp12(x, iso1, iso2, res1, res2, free = F)
}
\arguments{
\item{x}{A dtoto data.frame of the mixed species containing Ytot, densite1, densite2, Yesp1, Yesp2, nbplt and surfsolref vectors}

\item{iso1}{A dtoto data.frame of the isolated plants for Esp1 containing Ytot, densite1, densite2, Yesp1, Yesp2, nbplt and surfsolref vectors}

\item{iso2}{A dtoto data.frame of the isolated plants for Esp2 containing Ytot, densite1, densite2, Yesp1, Yesp2, nbplt and surfsolref vectors}

\item{res1}{A model fit for response to density of Esp1 with "a","b" and "beta" values}

\item{res2}{A model fit for response to density of Esp2 with "a","b" and "beta" values}

\item{free}{An optional parameter to force beta parameters as estimated from pure species}
}
\value{
A model fit result "model1", "model2", "parameters1", "parameters2"
}
\description{
Calculate Gamma competition coefficient for inter-specific competition
}
\examples{
dtoto <- tabtoto_compet
sp_dtoto <- split(dtoto, dtoto$keysc)
key <-"88-1 Fix2-nonFixSimTest Lusignan30IrrN2 -"
keypur <- "88-1 nonFixSimTest-nonFixSimTest Lusignan30IrrN2 -"
dat <- rbind(sp_dtoto[[key]], sp_dtoto[[keypur]])
pur1 <- dat[dat$densite2 == 0., ]
pur2 <- dat[dat$densite1 == 0., ]
diag <- dat[dat$densite == 400., ]
iso1 <- pur1[pur1$densite ==min(pur1$densite) , ]
iso2 <- pur2[pur2$densite ==min(pur2$densite) , ]
res1 <- Calc_Beta_coeff(pur1)
res2 <- Calc_Beta_coeff(pur2)
Calc_Gamma_coeffesp12(diag, iso1, iso2, res1, res2)
# sans forcage
Calc_Gamma_coeffesp12(diag, iso1, iso2, res1, res2, free=T)
}
