% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mef.R
\name{calc_neighb_param}
\alias{calc_neighb_param}
\title{This function ...}
\usage{
calc_neighb_param(
  tabpar,
  lspar,
  ls_idvois,
  ls_idKin,
  ls_idnonKin,
  cote = 16,
  nblignes = 16
)
}
\description{
This function ...
}
