% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_indices.R
\name{Calc_Rij_coefficients}
\alias{Calc_Rij_coefficients}
\title{Calculate Rij coefficients}
\usage{
Calc_Rij_coefficients(res1, res2, parameters1, parameters2)
}
\description{
Calculate Rij coefficients
}
