% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_indices.R
\name{Calc_Sij_coefficients}
\alias{Calc_Sij_coefficients}
\title{Calculate Sij coefficients}
\usage{
Calc_Sij_coefficients(res1, res2, parameters1, parameters2)
}
\description{
Calculate Sij coefficients
}
