% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_indices.R
\name{CalcOpt}
\alias{CalcOpt}
\title{Calculate density for Optimal overyiending in De Wit's simulation design + other indices}
\usage{
CalcOpt(modeltot, xx, yy)
}
\description{
Calculate density for Optimal overyiending in De Wit's simulation design + other indices
}
