% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_indices.R
\name{CalcPropactu50}
\alias{CalcPropactu50}
\title{Calculate sowing proportion achieving 50/50 in De Wit's simulation design + other indices}
\usage{
CalcPropactu50(modelesp1, modelesp2, idopt)
}
\description{
Calculate sowing proportion achieving 50/50 in De Wit's simulation design + other indices
}
