% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_plots.R
\name{Concat_ggplot_layers}
\alias{Concat_ggplot_layers}
\title{Add different ggplot layers into the same plot}
\usage{
Concat_ggplot_layers(ls_pltsim, layertype = "GeomLine")
}
\description{
Add different ggplot layers into the same plot
}
