% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_simtoto.R
\name{efficiencyCoucheney}
\alias{efficiencyCoucheney}
\title{This function returns the modelling Efficiency}
\usage{
efficiencyCoucheney(O, P)
}
\arguments{
\item{O}{observed values}

\item{P}{simulated values}
}
\value{
Modelling Efficiency (EF)
}
\description{
This function returns the modelling Efficiency
}
