% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simtoto.R
\name{gg_addplotobs}
\alias{gg_addplotobs}
\title{This function adds the observed points to plot_var}
\usage{
gg_addplotobs(plot_var, var_, obsOK, obsMerge)
}
\arguments{
\item{plot_var}{plot of the dynamics of the output variables of L-egume}

\item{var_}{Simulated variable}

\item{obsOK}{observation values of the same DOY range of the simulation values}

\item{obsMerge}{data frame containing the simulation and observation values}
}
\value{
A 2D graph
}
\description{
This function adds the observed points to plot_var
}
