% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_plots.R
\name{gg_plotObsSim}
\alias{gg_plotObsSim}
\title{Plotting simulation outputs versus observations}
\usage{
gg_plotObsSim(obssim, var_, name = "", colpt = "red")
}
\description{
Plotting simulation outputs versus observations
}
\examples{
ltoto <- ltoto_exemple
simmoy <- build_simmoy(ltoto, lsusm=names(ltoto))

# resize obs data.frame to the dim of simmoy
obsOK <- obs_exemple[obs_exemple$DOY \%in\% simmoy$STEPS,]
obsMerge <- simmoy[,c("STEPS","TT")]
names(obsMerge)[1] <- "DOY"
obsMerge <- merge(obsMerge, obsOK, by="DOY",all=T)

#corresponding names in sim/obs data.frames
corresp <- data.frame(sim=c("NBI","LAI","MSA"), obs=c("NBI-quart","surf_tot","MSaerien"))

#plot
var_ <- "NBI"
nomvarobs <- as.character(corresp[corresp$sim==var_,c("obs")])
obssim <- na.omit(data.frame(obs=simmoy[,var_], sim=obsMerge[,nomvarobs]))
gg_plotObsSim(obssim, var_, name="test")

}
