% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simtoto.R
\name{gg_plotsim}
\alias{gg_plotsim}
\title{This functions plots the dynamics of the output variables of L-egume}
\usage{
gg_plotsim(varsim, simmoy, simsd, name = " ", col = "blue", colref = "red")
}
\arguments{
\item{varsim}{simulated variable}

\item{simmoy}{output of build_simmoy function (mean values)}

\item{simsd}{output of build_simmoy function (standard deviation)}

\item{name}{name of the plot}

\item{col}{color of the plot}

\item{colref}{color of the second plot of the same graph}
}
\value{
A 2D graph
}
\description{
This functions plots the dynamics of the output variables of L-egume
}
