% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simtoto.R
\name{gg_plotsimVar}
\alias{gg_plotsimVar}
\title{This functions plots the dynamics of two output variables of L-egume}
\usage{
gg_plotsimVar(varsim, simmoy, simmoy_, simsd, simsd_)
}
\arguments{
\item{varsim}{simulated variable}

\item{simmoy}{output of build_simmoy function (mean values) for version 1}

\item{simmoy_}{output of build_simmoy function (mean values) for version 2}

\item{simsd}{output of build_simmoy function (standard deviation) for version 1}

\item{simsd_}{output of build_simmoy function (standard deviation) for version 2}
}
\value{
A 2D graph
}
\description{
This functions plots the dynamics of two output variables of L-egume
}
