% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mef.R
\name{init_dtoto_binary_zip1}
\alias{init_dtoto_binary_zip1}
\title{This function ...initiates dtoto colums from a list of zip file names according to names nomenclature with sd in run_legume_usm}
\usage{
init_dtoto_binary_zip1(ls_zip)
}
\description{
This function ...initiates dtoto colums from a list of zip file names according to names nomenclature with sd in run_legume_usm
}
