% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\docType{data}
\name{management}
\alias{management}
\title{Model Inputs}
\format{
A data frame with 367 rows and 10 variables
}
\usage{
data(management)
}
\description{
Mnagement input example
}
\details{
\itemize{
\item{Date:} {Date mm/dd/yy}
\item{year:} {year}
\item{month:} {month}
\item{day:} {day}
\item{DOY:} {Day of the Year}
\item{Coupe:} {Days of the cuts}
\item{Irrig:} {Irrigation}
\item{FertNO3:} {NO3 Fertilization, in kg/ha}
\item{FertNH4:} {NH4 Fertilization, in kg/ha}
\item{Hcut:} {Height of the cut, in cm}
}
}
\keyword{dataset}
\keyword{input}
