% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moy_simtoto.R
\name{moysimval}
\alias{moysimval}
\title{This function determines the average of a variable 'var' for all the plants in a canopy}
\usage{
moysimval(ltoto, lsusm, var, esp = NA, optSD = F)
}
\arguments{
\item{ltoto}{A large list of three data frames (seeds 0, 1 and 2)}

\item{lsusm}{names(ltoto)}

\item{var}{Output variable of L-egume}

\item{esp}{Species}

\item{optSD}{Optional parameter.Returns standard deviation of the sum of the individuals}
}
\value{
The mean value of 'var' for each DOY
}
\description{
This function determines the average of a variable 'var' for all the plants in a canopy
}
