% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regular_plots.R
\name{Plot2D_soilVar}
\alias{Plot2D_soilVar}
\title{Plot 2D spatio-temporal change of soil variables from the outHR file}
\usage{
Plot2D_soilVar(
  dat,
  var_ = "FTSW",
  epcouche = 5,
  Boundarycols = c("blue", "red"),
  vDOYs = NULL
)
}
\arguments{
\item{dat}{A dataframe with outHR file data}

\item{var_}{Name of the soil variable to plot}

\item{epcouche}{Soil voxel (cm)}

\item{Boundarycols}{A vector of two color names defining the extremes of the gradient of values}

\item{vDOYs}{A vector of DOY (Day of The Year) value to keep}
}
\value{
2D spatio-temporal plot of soil variables from the outHR file
}
\description{
Plot 2D spatio-temporal change of soil variables from the outHR file
}
\examples{
dat <- outHR_0
Plot2D_soilVar(dat, var_='FTSW')
Plot2D_soilVar(dat, var_='HRp')
Plot2D_soilVar(dat, var_='m_NO3', Boundarycols =c("brown", "green"))
Plot2D_soilVar(dat, var_='m_NN4', Boundarycols =c("brown", "green"))
Plot2D_soilVar(dat, var_='m_NO3', Boundarycols =c("brown", "green"), vDOYs=100:200)

}
