% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regular_plots.R
\name{Plot_OneOne_prop}
\alias{Plot_OneOne_prop}
\title{Plot two species response to density on the one-one diagonal for species proportions}
\usage{
Plot_OneOne_prop(x, parameters1, parameters2, dmax = 400, ...)
}
\arguments{
\item{x}{A dtoto data.frame of the mixed species containing Ytot, densite1, densite2, Yesp1, Yesp2, nbplt and surfsolref vectors}

\item{parameters1}{The result of a fit with Calc_Beta_coeff function}

\item{parameters2}{The result of a fit with Calc_Beta_coeff function}

\item{dmax}{Density}
}
\value{
A plot of two species response to density with its fit on the one-one diagonal at various densities for species proportions
}
\description{
Plot two species response to density on the one-one diagonal for species proportions
}
\examples{
dtoto <- tabtoto_compet
sp_dtoto <- split(dtoto, dtoto$keysc)
key <-"88-1 Fix2-nonFixSimTest Lusignan30IrrN2 -"
keypur <- "88-1 nonFixSimTest-nonFixSimTest Lusignan30IrrN2 -"
dat <- rbind(sp_dtoto[[key]], sp_dtoto[[keypur]])
pur1 <- dat[dat$densite2 == 0., ]
pur2 <- dat[dat$densite1 == 0., ]
diag <- dat[dat$densite == 400., ]
iso1 <- pur1[pur1$densite ==min(pur1$densite) , ]
iso2 <- pur2[pur2$densite ==min(pur2$densite) , ]
res1 <- Calc_Beta_coeff(pur1)
res2 <- Calc_Beta_coeff(pur2)
resg <- Calc_Gamma_coeffesp12(one_one, iso1, iso2, res1, res2)
parameters1 <- resg[["parameters1"]]
parameters2 <- resg[["parameters2"]]
Plot_OneOne_Resp_dtot(one_one, parameters1, parameters2, dmax=400.)
Plot_OneOne_prop(one_one, parameters1, parameters2, dmax=400.)

}
