% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regular_plots.R
\name{plot_ranking_lines}
\alias{plot_ranking_lines}
\title{Plot biomass trajetories}
\usage{
plot_ranking_lines(
  tab,
  id_refdec = 1,
  decHaut = 9,
  decBas = 2,
  id_refline = NA,
  titre = "",
  ymax = 1,
  xlab = "",
  ylab = ""
)
}
\description{
Plot biomass trajetories
}
