% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regular_plots.R
\name{Plot_resCE_SE}
\alias{Plot_resCE_SE}
\title{Plot Loreau's CEi and SEi coefficient for a diagonal}
\usage{
Plot_resCE_SE(resCE_SE, ...)
}
\description{
Plot Loreau's CEi and SEi coefficient for a diagonal
}
