% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regular_plots.R
\name{PlotMapVal}
\alias{PlotMapVal}
\title{Plot Map of values}
\usage{
PlotMapVal(
  tab,
  Val_param,
  cexfactor = 4,
  append = F,
  reverse = F,
  norm = NA,
  ...
)
}
\description{
Plot Map of values
}
