% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regular_plots.R
\name{Plt_Yresp_densite1}
\alias{Plt_Yresp_densite1}
\title{Plot single species response to density}
\usage{
Plt_Yresp_densite1(x, res, ...)
}
\arguments{
\item{x}{A dtoto data.frame of the pure species containing Ytot, densite, nbplt and surfsolref vectors}

\item{res}{The result of a fit with Calc_Beta_coeff function}
}
\value{
A plot of single species response to density with its fit
}
\description{
Plot single species response to density
}
\examples{
dtoto <- tabtoto_compet
sp_dtoto <- split(dtoto, dtoto$keysc)
key <-"88-1 Fix2-nonFixSimTest Lusignan30IrrN2 -"
keypur <- "88-1 nonFixSimTest-nonFixSimTest Lusignan30IrrN2 -"
dat <- rbind(sp_dtoto[[key]], sp_dtoto[[keypur]])
pur1 <- dat[dat$densite2 == 0., ]
res1 <- Calc_Beta_coeff(pur1)
Plt_Yresp_densite1(pur1, res1, main="", xlab="densite 1", ylab="Ytot")
}
