% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_simtoto.R
\name{pRMSEs}
\alias{pRMSEs}
\title{This function returns the percentage of the systematic error of the model}
\usage{
pRMSEs(rmse_, rmses_)
}
\arguments{
\item{rmse_}{Relative Mean Square Error}

\item{rmses_}{systematic error of the model}
}
\value{
percentage of the systematic error
}
\description{
This function returns the percentage of the systematic error of the model
}
