% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_simtoto.R
\name{pRMSEu}
\alias{pRMSEu}
\title{This function returns the percentage of the systematic error of the model}
\usage{
pRMSEu(rmse_, rmseu_)
}
\arguments{
\item{rmse_}{Relative Mean Square Error}

\item{rmseu_}{unsystematic error of the model}
}
\value{
percentage of the unsystematic error
}
\description{
This function returns the percentage of the systematic error of the model
}
