% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regular_plots.R
\name{QNtotvsProp}
\alias{QNtotvsProp}
\title{Plot of N-Overyielding against species proportion}
\usage{
QNtotvsProp(
  tabmoy,
  Ymax = 100,
  nom = "",
  optProp = "sowing",
  visuplot = T,
  visutext = T,
  supindices = T,
  labx = NA,
  ...
)
}
\description{
Plot of N-Overyielding against species proportion
}
