% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mef.R
\name{read_lsSD_MStot}
\alias{read_lsSD_MStot}
\title{This function reads the content of the simulation file "toto" and "paramSD" of a list of USM and stocks "ls_tabSD" and "ls_MStot"}
\usage{
read_lsSD_MStot(ltoto, lparamSD, param_name = "Len")
}
\description{
This function reads the content of the simulation file "toto" and "paramSD" of a list of USM and stocks "ls_tabSD" and "ls_MStot"
}
