% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_totosim.R
\name{read_ltoto}
\alias{read_ltoto}
\title{This function takes the group of USMs of a specific Key, reads the content of the simulation file "toto" of each USM and stocks the output in a list}
\usage{
read_ltoto(ls_toto)
}
\arguments{
\item{ls_toto}{list of toto files belonging to the same key}
}
\value{
A list. Each element of this list correspond to a single USM
}
\description{
This function takes the group of USMs of a specific Key, reads the content of the simulation file "toto" of each USM and stocks the output in a list
}
