% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_model.R
\name{rmse}
\alias{rmse}
\title{Calculate the Root Mean Square Error (RMSE) of predicted model outputs}
\usage{
rmse(O, P)
}
\arguments{
\item{O}{An observed vector of values}

\item{P}{A vector of values predicted by a model}
}
\value{
The Root Mean Square Error between \code{O} and \code{P}
}
\description{
Calculate the Root Mean Square Error (RMSE) of predicted model outputs
}
\examples{
obs <- c(1, 2, 3, 5)
sim <- c(0.9, 2.2, 3.1, 5)
rmse(obs, sim)
}
