% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_simtoto.R
\name{rmsesCoucheney}
\alias{rmsesCoucheney}
\title{This function returns the systematic error of the model}
\usage{
rmsesCoucheney(O, P)
}
\arguments{
\item{O}{observed values}

\item{P}{simulated values}
}
\value{
systematic error RMSEs
}
\description{
This function returns the systematic error of the model
}
