% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_simtoto.R
\name{rmseuCoucheney}
\alias{rmseuCoucheney}
\title{This function returns the unsystematic error of the model}
\usage{
rmseuCoucheney(O, P)
}
\arguments{
\item{O}{observed values}

\item{P}{simulated values}
}
\value{
unsystematic error RMSEu
}
\description{
This function returns the unsystematic error of the model
}
