% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_simtoto.R
\name{rrmseCoucheney}
\alias{rrmseCoucheney}
\title{This function returns the relative mean square error in \%}
\usage{
rrmseCoucheney(O, P)
}
\arguments{
\item{O}{observed values}

\item{P}{simulated values}
}
\value{
Relative Mean Square Error (RMSE)
}
\description{
This function returns the relative mean square error in \%
}
