% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toto.R
\docType{data}
\name{toto}
\alias{toto}
\title{Model Simulations}
\format{
A data frame with 53940 rows and 10 variables
}
\usage{
data(toto)
}
\description{
An output example
}
\details{
\itemize{
\item{TT} {Thermal time, in degree.days}
\item{time} {no unit}
\item{cutNB} {no unit}
\item{SurfPlante} {Total plant leaf area, in \ifelse{html}{\out{m<sup>2</sup>}}{\eqn{m^2}}}
\item{PARiPlante} {Plant intercepted PAR, in J.PAR}
\item{epsi} {no unit}
\item{dMSaer} {no unit}
\item{Hplante} {Canopy height, in cm}
\item{Dplante} {no unit}
\item{RLTot} {no unit}
\item{RDepth} {Root depth, in m}
\item{MSaerien} {Total plant shoot dry biomass, in g}
\item{MSfeuil} {no unit}
\item{MStot} {no unit}
\item{countSh} {no unit}
\item{countShExp} {no unit}
\item{demandC} {no unit}
\item{Leaf_Stem} {no unit}
\item{NBsh} {Number of primary shoots per plant}
\item{NBI} {Number of phytomers per primary axes of plant}
\item{NBD1} {no unit}
\item{NBB} {no unit}
\item{FTSW} {Fraction of transpirable soil water}
\item{Etransp} {no unit}
\item{DemandC_Feuil} {The demand of leaves for carbon}
\item{DemandN_Pet} {The demand of petioles for nitrogen}
\item{DemandN_Stem} {The demand of stems for nitrogen}
\item{DemandN_Tot} {The demand of individual plants for nitrogen}
\item{DemandN_Tot_Aer} {The demand of aerial parts of individual plants for nitrogen}
\item{Npc_aer} {no unit}
\item{NNI} {Plant nitrogen nutrition index}
\item{Ndfa} {Proportion of N derived from fixation}
\item{Qfix} {no unit}
\item{Naerien} {no unit}
\item{Nuptake_sol} {no unit}
\item{R_DemandC_Root} {no unit}
\item{SRL} {no unit}
\item{dMSenFeuil} {no unit}
\item{dMSenTige} {no unit}
\item{MS_pivot} {no unit}
\item{MS_rac_fine} {no unit}
\item{R_DemandC_Shoot} {no unit}
\item{RUEpot} {no unit}
\item{RUE} {Radiation use efficiency, in \ifelse{html}{\out{g.J.PAR<sup>-1</sup>}}{\eqn{g.J.PAR^-1}}}
\item{Npc_piv} {no unit}
\item{Npc_rac_fine} {no unit}
\item{dRLenSentot} {no unit}
\item{dMSenRoot} {no unit}
\item{RLTotNet} {no unit}
\item{MS_rac_fineNet} {no unit}
\item{perteN_rac_fine} {no unit}
\item{NBphyto} {no unit}
\item{NBapexAct} {no unit}
\item{transpi} {no unit}
\item{cumtranspi} {no unit}
\item{aliveB} {no unit}
\item{TTphyllo} {no unit}
\item{dTT} {no unit}
\item{Udev} {no unit}
\item{Udevstress} {no unit}
}
}
\keyword{datasets}
