% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mef.R
\name{Which_decile}
\alias{Which_decile}
\title{This function determines to which decile  is a value in a vector of values}
\usage{
Which_decile(valparams)
}
\description{
This function determines to which decile  is a value in a vector of values
}
