% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_indices.R
\name{Yresp_densite1}
\alias{Yresp_densite1}
\title{Single species response to density}
\usage{
Yresp_densite1(a, b, densite)
}
\arguments{
\item{a}{A parameter..}

\item{b}{A parameter..}

\item{densite}{plant density..}
}
\value{
Yield at density
}
\description{
Single species response to density
}
