% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_indices.R
\name{Yresp_densite2_inv}
\alias{Yresp_densite2_inv}
\title{Two species response to density with inverse calculation}
\usage{
Yresp_densite2_inv(a, beta, gamma, densite1, densite2)
}
\arguments{
\item{a}{A parameter..}

\item{beta}{A parameter..}

\item{gamma}{A parameter..}

\item{densite1}{plant density..}

\item{densite2}{plant density..}
}
\value{
inverse of Yield of Esp1 according to densities of Esp1 and Esp2 and intra- / inter- specific competition coefficients
}
\description{
Two species response to density with inverse calculation
}
