% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regular_plots.R
\name{YtotvsProp}
\alias{YtotvsProp}
\title{Plot of Overyielding against species proportion}
\usage{
YtotvsProp(
  tabmoy,
  Ymax = 2200,
  nom = "",
  optProp = "sowing",
  visuplot = T,
  visutext = T,
  supindices = T,
  labx = NA,
  col2 = 2,
  ...
)
}
\description{
Plot of Overyielding against species proportion
}
