% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_totosim.R
\name{build_dtoto}
\alias{build_dtoto}
\title{This function builds dtoto: a data frame grouping all the IDusms simulated by the model}
\usage{
build_dtoto(ls_toto)
}
\arguments{
\item{ls_toto}{list of all toto files of the same experiment (.csv files)}
}
\value{
A data frame. Each row of this data frame corresponds to an IDusm
}
\description{
This function builds dtoto: a data frame grouping all the IDusms simulated by the model
}
