% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mef.R
\name{build_simmoy}
\alias{build_simmoy}
\title{Calculate the the average or standard deviation of a series of simulations for a set of predefined variables}
\usage{
build_simmoy(ltoto, lsusm, esp = NA, optSD = F)
}
\arguments{
\item{ltoto}{A list of data frames. Each element of this list correspond to the simulation output of a single USM}

\item{lsusm}{The list of USMs to consider in ltoto}

\item{esp}{An optional parameter to specify the name of a species to consider in case of multi-species simulations}

\item{optSD}{An optional parameter to specify if standard deviation should be calculated instead of mean}
}
\value{
A simmoy data.frame with variables : STEPS, TT, NBI, NBphyto, LAI, MSA, MSArec, MSAnonrec, MSpiv, MSracfine, MSrac, RDepth, Hmax, FTSW, NNI, R_DemandC_Root, cutNB, Npc_aer, Ndfa, Epsi, NBsh
}
\description{
Calculate the the average or standard deviation of a series of simulations for a set of predefined variables
}
\examples{
ltoto <- ltoto_exemple
simmoy <- build_simmoy(ltoto, lsusm=names(ltoto))
plot(simmoy$STEPS, simmoy$NBI, type='l')
}
