% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_model.R
\name{efficiencyCoucheney}
\alias{efficiencyCoucheney}
\title{Calculate the modelling Efficicency (EF) of predicted model outputs}
\usage{
efficiencyCoucheney(O, P)
}
\arguments{
\item{O}{An observed vector of values}

\item{P}{A vector of values predicted by a model}
}
\value{
The modelling Efficicency of \code{P} with respect to \code{O}
}
\description{
Calculate the modelling Efficicency (EF) of predicted model outputs
}
\examples{
obs <- c(1, 2, 3, 5)
sim <- c(0.9, 2.2, 3.1, 5)
efficiencyCoucheney(obs, sim)
}
