% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_plots.R
\name{gg_addplotobs}
\alias{gg_addplotobs}
\title{Add observed points to an existing plot of simulation (plotsim)}
\usage{
gg_addplotobs(plot_var, var_, obsOK, corresp, colobs = "red")
}
\description{
Add observed points to an existing plot of simulation (plotsim)
}
\examples{
ltoto <- ltoto_exemple
simmoy <- build_simmoy(ltoto, lsusm=names(ltoto))
simsd <- build_simmoy(ltoto, lsusm=names(ltoto), optSD = T)
plt_test <- gg_plotsim("NBI", simmoy, simsd, name="test")

# resize obs data.frame to the dim of simmoy
obsOK <- obs_exemple[obs_exemple$DOY \%in\% simmoy$STEPS,]
obsMerge <- simmoy[,c("STEPS","TT")]
names(obsMerge)[1] <- "DOY"
obsMerge <- merge(obsMerge, obsOK, by="DOY",all=T)

#corresponding names in sim/obs data.frames
corresp <- data.frame(sim=c("NBI","LAI","MSA"), obs=c("NBI-quart","surf_tot","MSaerien"))

#add obs points to plot
plt_test <- gg_addplotobs(plt_test, "NBI", obsMerge, corresp)
plt_test
}
