% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_plots.R
\name{gg_plotsim}
\alias{gg_plotsim}
\title{Plotting simulation outputs}
\usage{
gg_plotsim(varsim, simmoy, simsd, name = "", col = "blue", colref = "red")
}
\arguments{
\item{varsim}{name of the output variable}

\item{simmoy}{data frame containing the average simulated outputs of all the output variables}

\item{simsd}{data frame containing the standard deviation of simulated outputs of all the output variables}

\item{name}{Name...}

\item{col}{Color...}

\item{colref}{Color...}
}
\value{
A graph showing the dynamics of the selected output variable
}
\description{
Plotting simulation outputs
}
\examples{
ltoto <- ltoto_exemple
simmoy <- build_simmoy(ltoto, lsusm=names(ltoto))
simsd <- build_simmoy(ltoto, lsusm=names(ltoto), optSD = T)
gg_plotsim("LAI", simmoy, simsd, name="test")
}
