% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mef.R
\name{moysimval}
\alias{moysimval}
\title{Calculate the the average or standard deviation of a series of simulations}
\usage{
moysimval(ltoto, lsusm, var, esp = NA, optSD = F)
}
\arguments{
\item{ltoto}{A list of data frames. Each element of this list correspond to the dynamic simulation output of a single USM}

\item{lsusm}{The list of USMs to consider in ltoto}

\item{var}{Name of the variable to consider for calculation}

\item{esp}{An optional parameter to specify the name of a species to consider in case of multi-species simulations}

\item{optSD}{An optional parameter to specify if standard deviation should be calculated instead of mean}
}
\value{
A vector of average or stand deviation of the variable 'var'
}
\description{
Calculate the the average or standard deviation of a series of simulations
}
\examples{
ltoto <- ltoto_exemple
NBI <- moysimval(ltoto, lsusm=names(ltoto), var='NBI')
sdNBI <- moysimval(ltoto, lsusm=names(ltoto), var='NBI', optSD=T)
plot(NBI, type='l')
# add standard deviations
segments(1:length(NBI), NBI, 1:length(NBI), NBI+sdNBI, col=2)
segments(1:length(NBI), NBI, 1:length(NBI), NBI-sdNBI, col=2)
}
