% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_model.R
\name{pRMSEu}
\alias{pRMSEu}
\title{Calculate the partial unsystematic component of Root Mean Square Error (pRMSEu) of predicted model outputs}
\usage{
pRMSEu(rmse_val, rmseu_val)
}
\arguments{
\item{O}{An observed vector of values}

\item{P}{A vector of values predicted by a model}
}
\value{
The partial unsystematic Root Mean Square Error between \code{O} and \code{P}
}
\description{
Calculate the partial unsystematic component of Root Mean Square Error (pRMSEu) of predicted model outputs
}
\examples{
obs <- c(1, 2, 3, 5)
sim <- c(0.9, 2.2, 3.1, 5)
rmse_val <- rmse(obs, sim)
rmses_val <- rmsesCoucheney(obs, sim)
rmseu_val <- rmseuCoucheney(obs, sim)
pRMSEs(rmse_val, rmses_val)
pRMSEu(rmse_val, rmseu_val)
pRMSEs + pRMSEu
}
