% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mef.R
\name{read_ltoto}
\alias{read_ltoto}
\title{This function reads the content of the simulation file "toto" of a series of USM and stocks the dynamic daily output into a list}
\usage{
read_ltoto(ls_toto)
}
\arguments{
\item{ls_toto}{A list of toto files paths}
}
\value{
A list of data frames. Each element of this list correspond to the simulation output of a single USM
}
\description{
This function reads the content of the simulation file "toto" of a series of USM and stocks the dynamic daily output into a list
}
\examples{
path_ltoto1 <- system.file("extdata", "toto_6129_l-egume_Orca-Orca_homogeneous0_scenario12-12_DigitLuz10LD_0_DigitLuz10_-_.csv", package = "RPackVGL")
path_ltoto2 <- system.file("extdata", "toto_6130_l-egume_Orca-Orca_homogeneous0_scenario12-12_DigitLuz10LD_1_DigitLuz10_-_.csv", package = "RPackVGL")
path_ltoto3 <- system.file("extdata", "toto_6131_l-egume_Orca-Orca_homogeneous0_scenario12-12_DigitLuz10LD_2_DigitLuz10_-_.csv", package = "RPackVGL")

ls_toto <- c(path_ltoto1, path_ltoto2, path_ltoto3)
my_ltoto <- read_ltoto(ls_toto)
}
