% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_model.R
\name{rmsesCoucheney}
\alias{rmsesCoucheney}
\title{Calculate the systematic component of Root Mean Square Error (RMSEs) of predicted model outputs}
\usage{
rmsesCoucheney(O, P)
}
\arguments{
\item{O}{An observed vector of values}

\item{P}{A vector of values predicted by a model}
}
\value{
The systematic Root Mean Square Error between \code{O} and \code{P}
}
\description{
Calculate the systematic component of Root Mean Square Error (RMSEs) of predicted model outputs
}
\examples{
obs <- c(1, 2, 3, 5)
sim <- c(0.9, 2.2, 3.1, 5)
rmsesCoucheney(obs, sim)
}
\seealso{
Other rmse: 
\code{\link{rmseuCoucheney}()}
}
\concept{rmse}
