% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_model.R
\name{rrmseCoucheney}
\alias{rrmseCoucheney}
\title{Calculate the relative Root Mean Square Error (rRMSE) of predicted model outputs}
\usage{
rrmseCoucheney(O, P)
}
\arguments{
\item{O}{An observed vector of values}

\item{P}{A vector of values predicted by a model}
}
\value{
The relative Root Mean Square Error between \code{O} and \code{P} to the mean of observed values
}
\description{
Calculate the relative Root Mean Square Error (rRMSE) of predicted model outputs
}
\examples{
obs <- c(1, 2, 3, 5)
sim <- c(0.9, 2.2, 3.1, 5)
rrmseCoucheney(obs, sim)
}
